+++ engine/src/geometry/tms_geometry.f90
@@ -1,5 +1,6 @@
 module tms_geometry
   use tms_kinds, only : dp
+  use tms_constants, only : pi
   implicit none
   private
 
@@ -34,5 +35,29 @@ module tms_geometry
     type(rubber_geometry_t) :: rubber
     type(inertia_ring_geometry_t) :: inertia_ring
   end type tvd_geometry_t
+
+  public :: calculate_rubber_polar_area_moment
+
+contains
+
+  !> Annüler elastomer kesitin polar geometrik alan momentini hesaplar.
+  !!
+  !! Fiziksel açıklama: Polar alan momenti, kesit geometrisinin burulmaya
+  !! karşı rijitlik katkısını temsil eder; polar kütle ataleti değildir.
+  !! Matematiksel açıklama: Jp = pi/2 * (ro^4 - ri^4).
+  !! Girdi: rubber iç ve dış yarıçapları metre (m) cinsindedir. Eksenel
+  !! uzunluk bu geometrik kesit hesabında kullanılmaz.
+  !! Çıktı: Polar alan momenti metrenin dördüncü kuvveti (m^4) cinsindendir.
+  !! Varsayımlar ve geçerlilik: Kesit eksenel simetrik ve annülerdir;
+  !! 0 <= ri < ro olmalıdır. Yordam bu önkoşulları doğrulamaz.
+  !! Ayrıntılar: docs/mathematics/torsional-physics-core.md.
+  pure elemental function calculate_rubber_polar_area_moment(rubber) &
+      result(polar_area_moment_m4)
+    type(rubber_geometry_t), intent(in) :: rubber
+    real(dp) :: polar_area_moment_m4
+
+    polar_area_moment_m4 = 0.5_dp * pi * &
+      (rubber%outer_radius_m**4 - rubber%inner_radius_m**4)
+  end function calculate_rubber_polar_area_moment
 
 end module tms_geometry